Allow container to run as a non-root user
This resolves the issue of running containers as a non-root user. By
setting the user to 1000 we are guarenteed to start Couchbase as the
couchbase user. In Openshift deployments we can also run as a random
user (the USER command is ignored) by giving our Couchbase
installation and runit the ability to be executed and run by any user.